move `get_logger` to another file as it causes circular imports if used in `databooks.data_models.notebook`
